Fix crash when `Save As` is triggered by a `Save`
If the user created a new file, then tried to `Save` the file, it detected that
it hadn't been saved before, and so called the `Save As` routine.

Both the `Save` and `Save As` (and incidentally `Save As Copy`) routines
temporarily disabled the autosave feature using the Perl `local` keyword,
which stores a backup of a global variable, restoring it when the current
block is exited.

It appears that this mechanism in Perl does not work if a global variable is
saved, then in a subroutine called from within the block (i.e. while the `local`
is still in effect) the global variable is `local`ed again. Perl crashed when
the inner block was exited.

Recoded to backup to a `my` scope variable, with a few adjustments to
ensure that the restoring lines couldn't be skipped by early exit of the
routines.

Fixes #960