Add `Rejoin Rows` to Table Effects dialog
Some OCRed files have each cell of a table on a new line, with a blank line
indicating the end of a row.

To support PMs prepping files from these OCRed files, add a button to rejoin
these split rows, so that each cell is separated by 2 spaces from the next cell,
and all cells from the same row appear on the same line of text, which is what
would be expected output from the rounds.