Make spell query user dict only contain added words
Instead of the user global dictionary containing all the default words with the
user's additions appended, use the default dictionary to load the default
words and have the user dictionary just contain any words the user adds.

Advantages:
1. In a future release, we might want to ship an improved default dictionary,
which would be hard to manage under the previous scheme since the
dictionary being used contained a mixture of default and user-added words.
2. The user dictionary is small and less scary for manual editing.
3. Fits better with potential future use of `--home` to store user settings
outside the release.