Allow broader range of names for user-defined scannos folder
Due to `scannos` being a folder directly under the release (as opposed to
an executable in a subfolder of `tools`, like most of the other settings that
are configurable using `Set File Paths`, there are limitations on the name
a user can choose for their own customised scannos folder.

Previously, the user could not choose a folder name that ended in `scannos`,
e.g. `custom_scannos`, or the next time GG was run, the program reset it to
the default location under the release. The name `scannos_custom` was OK.

Slight adjustment to code now allows names like `custom_scannos`, or
`myscannos`. The only name that does not work is precisely `scannos`, since
it would be hard to know if that was under a release or not.

The manual already described the previous restriction and will be edited to
reflect the new, lesser restriction.