Highlight queried spellings more accurately
When attempting to highlight where the queried word was on the line, the
search found the first occurrence of the queried string, which might not be
the queried word, e.g. given `young 'un`, it would highlight the letters `un`
in the word `young`, rather than the whole word `un` that appears later.

Instead of trying to find the word on the line, new algorithm tracks the
column as each word is checked. This necessitates keeping the line's
alignment correct when replacing non-word characters, so each character
is replaced with one space. Seems to be more intuitive than the old method,
with just a minor tweak to the column when a leading apostrophe or
single quote is trimmed.

The old regex for detecting and removing block markup didn't cope with the
`/I[8.3,70]` case where margins are customized, so more complexity was
needed. In fact removing block markup wasn't needed since it consists of a
single letter, which is never flagged as a spelling query anyway. As only a
small percentage of lines contain block markup, it's quicker to just spell
check those, than to check every line to see it it has block markup.