Allow Spell Query to work with only a `user` dictionary
There is a `default` dictionary in the `data` folder, and a `user` dictionary which
may be in the `data` folder or the `GGprefs` folder.

Previously, the `default` had to exist, but now it is optional, as long as the `user`
dictionary exists. This allows users to add a language by putting a wordlist in
their GGprefs folder named `dict_LANG_user.txt` without needing to access
files under the release. There is no performance overhead of this change.