Avoid undefined variable when setting HTML attribute (#996)
When right-clicking (or ctrl-clicking) a button at the top of the HTML
markup dialog to configure it, a reference to an htmlentryattribhash
element gets passed to the dialog routine.
This brings it into existence, but initially it's undefined. If GG tries to write
the settings file while the dialog is popped (e.g. if GG loses & gains focus)
or if user cancels dialog, the undefined variable gives an error when
writing to the settings file.
Solved by initialising it to empty string if it's undefined.

Fixes #992