Simplify method for user-customized HTML header
Existing scheme requires the user to start GG for the first time,
which then causes the default header to be copied to the header
file, then edit the header file to customize it as they wish. When
they receive a new release, they either copy the header file to
the new release (or rely on the GGprefs scheme) both of which
mean they will not get the benefit of any changes to the default
header, unless they do some comparison/merging operation.

New scheme is a bit like user dictionaries. If the file `header_user.txt`
exists, then insert the default header followed by that user
header file. Since the user CSS will come after the default, it can
override it. Normally, there will be no need to do anything for a new
release, unless a change to the default header happens to clash
with the user's customization.

If `header_user.txt` doesn't exist, then behavior is unchanged.