Handle LOTE prefixes in Spell Query
Some languages use `l'`, `quest'`, etc., before a word.
If the "prefix" and the main word are both in the dictionary, then
accept the combined word with apostrophe.

The prefix as it appears in the dictionary may or may not have
the apostrophe - it's safer if it does, otherwise the abbreviated
prefix would also be accepted as a standalone word.