Remove jeebies' need for external `.jee` files
Instead of the jeebies executable loading `he.jee` and `be.jee`,
adjust the build process to include the data within the source:
process each `.jee` file to format it as initialization data for the
`char * []` variables used in the source; include these generated
`.inc` files in the C source.

There is now no need for the `.jee` files to be shipped with the
Windows version, nor for any data files to be moved into a `bin`
dir on other platforms. Once the executable is built by the
user, they can move that to a suitable `bin` dir to make it available
to all releases.