Fix jeebies Mac compiler warning
Since code triggering the "dangling else" warning did the same
in the `if` as in the `else`, removed that test.

This made the `isCR` variable pointless, so removed that too.

This left an empty `if` block, so swapped logic to put the `else`
block contents in the `if` block.

I believe this will mean no change in behavior - it appears to
give the same results as the old version.