Detect end-of-line he/be errors correctly
Previous PR #1012 broke the detection of he/be errors across
line boundaries. The replacement of `flgets` with `fgets` didn't
take account of the fact that `fgets` keeps the newline character.

Maybe I should have resisted the temptation to tidy up :)

Rather than process the line `fgets` returns, or alter the he/be
processing to cope with line endings, it seems safest and simplest
to go back to `flgets`