Add W3C epubcheck tool to Guiguts
We already bundle the HTML and CSS validators. The epubcheck
tool validates epub files. It is therefore slightly different to the
other tools in that the user has to select a file to be checked,
rather than it checking the currently loaded file.

Packaging seems to work fine when generating a release.