Shorten EPUBCheck error messages (#1024)
* Shorten EPUBCheck error messages

Error messages from EPUBCheck contain the full pathname of
the epub being checked, plus the internal path to the html file
inside the epub. The internal html file has a long name with about
20 digits plus the name of the epub plus an index number.
This can easily amount to 70+ characters before the error
message even gets started.

Trim all of this down to `...-N.html` so error message can be read
more easily. This contains sufficient information to find the
error in the epub using Calibre editor or by unzipping it. Will
be documented in the manual.

Also, hide the query count label since not needed for this check.

Fixes #1023