Generate non-serialized HTML5 by default
Use a global flag to control whether void elements such as `<br>`
are closed, i.e. `<br />`.
Default is not to close, as preferred by PG.
Global flag is saved in `setting.rc` which could be edited manually
to set to true. Not available in user interface though, to
discourage experimentation and confusion :)

When set to true, testsuite generates identical HTML to previous
version. When set to false, only differences are void element
closures.