Remove `Nu XHTML Checker` menu entry (#1029)
Since we are no longer producing XHTML, it makes no sense to have
the check available to the user.

Note that only the entry has been removed for now, to allow for it
to be restored along with XML serialization if there's another
change in PG's policy. In due course, it could be removed.

Also note that there is no "tool" to remove, because it use the
same tool as the HTML check, but with a `.xhtml` file extension.