Retrieve Nu validator by tag not release
Nu validator is not being released regularly, just tagged.
Git clone tagged version and build `vnu.jar`

Note for Windows development: requires Mingw64 shell since
Mingw32 shell cannot deal with a file created during the build
process with a Unicode character in the filename. When
attempting to tidy after the build, the `rm -rf` fails claiming
the `docs` dir is not empty, because `rm` can't delete a file
in the dir.