Stop double-clicking a word from including spaces
As part of #941 which made navigation, selecting and deleting
by word chunks more consistent, double-clicking a word
selected from the previous word start to the next one (similar
to MS Word).

This was unpopular with users, so wind this back to only
select the word itself, not any trailing spaces.