Improve extension of selection after word select
After word (or line) select using double/triple click, the anchor
for the selection was left at the click point (ie mid-word) which
led to weirdness when subsequently extending the selection
using Shift+arrow keys.

Now force the anchor to the start and the cursor to the end and
it behaves more logically.