Insert page break code if 4 blank lines precede block (#1050)
Four blank lines normally indicates a chapter break. If there is
regular text following 4 blank lines, the text is converted to an
`h2` chapter heading. However, especially in frontmatter, a chunk
of text, e.g. a dedication, might be on a separate page, with an
`hr` required before it in the HTML version, and a forced page
break in paged media.

This commit interprets 4 blank lines before most relevant block
markup (`/* /$ /r /c /p`) to mean that an `hr` should be added,
then a chapter div used to force a page break. Note that `/#`
is not supported because it can have other block types nested
inside it, which could cause chaos if a chapter div was opened
for the `/#` and then another one opened inside it for `/r` for
example. Also, `/#` is processed differently by the conversion
code. This limitation will be documented in the manual.