Add optional ruler/column numbers
Similar to the row numbers which can be displayed on the left of
the text window, this implements a column numbers "ruler" widget
across the top.

Similar configuration to the row numbers widget is available, but
not exposed to the user.

Turn on/off via Preferences menu or by Shift-right-clicking the
row/column display in the status bar.

Ruler is one-based rather than zero-based (as used internally), and
existing Highlight Alignment Column has been adjusted to match
the highlighted column number in the ruler.

Internal LineNumberText code has been simplified a little, removing
historical attempt to support plain Text widget rather than the GG
TextUnicode widget. Some comments also added.