Avoid `uninitialized` error exporting prep text (#1076)
If page markers have been corrupted, e.g. by editing outside GG,
code could fail to grab a page full of text to write to prep text file.
It's not practical to try to undo the corruption, so just warn the
user. It's no problem to a PM to delete the bin file, since at this
point the page separators are still in the file, so when they
restart, the page markers will be recalculated.