Allow user to prioritize display of footnote error types
If a footnote has two errors, e.g. it's a duplicate and a long footnote,
then only one of the errors can be shown by coloring the footnote in
the Footnote Check dialog list.

This commit allows the user to click a colored label at the top to
make that error type the highest priority, so all footnotes with that
error will be highlighted in the appropriate color.

Also a bit of code tidying to reduce duplication and put the link
between error types and colors in just one place.

Fixes #1072