Fix minor rewrap bugs (#1086)
1. Page markers on the blank line following `C/` or `R/` markup
caused extra blank line to be inserted.
2. Test for markup ending wasn't anchoring it to start of line, so
something like `i/` in the body of the text was being treated as
closure
3. Normally a blank line should follow `R/` or `C/`, but it is
permitted to omit, e.g. at end of illo caption, so next line is
just a `]`. This was then being right or center justified instead
remaining at left margin.
4. Bad markup, e.g. closing when not open, could cause global
variable to retain it's non-zero value when wrapping was complete.
This then affected the next attempt to wrap, even if the bad markup
was corrected.

Fixes #1085