Restore begin and end messages to SQ
Since the beginning/complete messages don't link to a place in
the file, the refresh code wasn't inserting them into the sorted
list.
1. Extract the text of those messages so it is not repeated in
the code multiple times
2. Specifically allow insertion of those messages
3. Comment the code to note that although the sorting code
works fine for SQ, there are other checks for which it wouldn't.
It's unlikely that such checks would ever require sorting, since
they don't have row:col numbers to sort by, but worth making
that explicit in the comments.
4. Minor safety check to ensure index to the active error message
does not get used if it is not valid.