Add block search options (#1089)
Three features to improve ability to search for blocks and then do
S&R operations within the block

1. Previously removed by #928 because only some block types
were implemented, and it took a lot of space in the Search menu.
No-one seemed to notice or care for a long time. Request for
its return has been implemented - as a submenu so as not to
clog Search menu - also improved to avoid code duplication,
to make it cope better when it doesn't find any more blocks,
and to cope better when switching from searching forwards to
backwards and vice versa. All block types now implemented,
ordered so that if menu is torn off, user can shrink it to just
show the most important ones.

2. There is a general "stay on top" flag already, and a specific flag
for the WF dialog. This implements a similar specific flag for the
S/R dialog.

3.  Add find any block-type to search submenu
Search for any line that begins with `/`
There may be very occasional lines where that happens in the
text, but rare enough and consequences trivial enough that the
benefit is worth the risk