Extend Find Match to handle block markup
If quote/bracket not found near cursor, code checked for HTML
tag. If no tag found, this commit also checks for block markup,
consisting of `/` preceded or followed by `*`, `#`, `$` or `[a-z]`.
This covers all block markup using symbols, and all current and
future markup using a letter.

This means that if the user selects `/H` (which is not
valid block markup at the moment), it will match with `H/`.
That is not a problem, since firstly the user will not have `/H`
in their file, and if they did, they wouldn't be trying to match
it with 'H/', and if they did try, it will work. It will also work for
new block types in future.

Also note that matching is case-sensitive, e.g. '/R` will not match
`r/`. This is a "feature" to encourage consistency, not an
unfortunate by-product of lazy programming :)

Fixes #1099