Adjust which match checks permit nesting (#1116)
The Unmatched item checks now report when some items are
nested. "Nested" means contained within an item of the identical
type, e.g. `/*` within `/*` would be reported, but not `/R` within `/*`.
Although this could lead to occasional false positives, this is better
than failing to detect badly formed items.
Specifically
1. Brackets may not be nested within their own type.
2. DP tags may not be nested
3. Only `/#` block markup may be nested
4. All non-void HTML tags may be nested. It may not be valid HTML
but that is not the purpose of the check. Primarily it is to detect
`<p>`, `<span>`, `<div>`, etc. that are not closed correctly.
5. The tag check is now split for DP and HTML tags to be treated
differently.