Remove duplicate text conversion buttons
In the Txt menu, several buttons were duplicates of buttons
in the Txt Conversion Palette. These related to conversion of
italic, bold, smallcap and thought-break markup. Because
these are typically clicked once per project, there is no need
for two versions of them.

The manual scarcely mentioned these duplicate options, but
just said the Palette (which is unchanged) is "the quickest and
easiest way of doing these conversions"

Fixes #1141