Remove unused rewrap parameters and code
There were already flags, fixes and tweaks in the rewrap code
to cope with what is wanted for Guiguts. Quite a lot of code
could never be executed. It's highly unlikely the GG use of
rewrap will be changed in the near future to take advantage
of those features, and if it is, an old version of ReflowGG can
always be dug up.

Rewrapping appears to behave identically to previously.

Fixes #1140