Use button/label font for toolbar
Buttons & labels use either the system default font, or
one that the user configures. The same font is now being
used for the toolbar for consistency.

Also, the large bold font used in the "Working" popup is
created as a named font (like the other fonts used) rather
than declared ad hoc in arguments to the Label creation
sub. This means it can be used of the same type of font
is needed elsewhere without duplicating the definition.

Fixes #1119