Save non-ASCII author name correctly
Specific example was a non-breaking space between initials,
but would have failed on any non-ASCII characters in author's
name. Problem arose because string is written out to `.bin` file
as perl code to be executed back in on file load, so higher
ordinal characters need escaping. Search/replace history strings
already had a mechanism when written to `setting.rc`, so
extracted and re-used that.

It would be fine to write every character out in hex, but it's
more readable to let the alphanumeric and spaces be
written out in native form, since they don't cause problems.

Fixes #1134