HTML-convert `<`, `>` that can't be part of a tag
When a text file is converted to HTML, angle brackets that are
part of the text (not tags) should be converted to HTML entities.
Previous code didn't convert text like `1<2` or `a>b`
Current code converts anything except a tag consisting of `<`
followed by any number of characters except angle brackets
then a closing `>`.  This is broader than tags that will typically
be found, but safer. It is better not to convert something
that should be (since the HTML validator should catch it) than
to convert something that shouldn't (since that may not
trigger the validator and could mean HTML tag code will
appear in the text.)

Fixes #1135