Make right-align shift left as well as right (#1161)
If a line in a right-align block extends beyond the right margin,
then consider if it's possible to move the block leftwards to
align with the right margin. This is possible if every line has
sufficient leading spaces in addition to those used in the left
margin (e.g. within block markup)

Only if the width of the block (from the leftmost non-space
on any line to the rightmost non-space on any line) is wider
than the gap between left and right margins do we have to
admit defeat.

Fixes #1133