Add `Convert DP diacriticals` feature
Within selected range(s), or whole file, convert DP diacritical
markup, e.g. `[=a]` or `[a.]` to the Unicode character(s) it
represents. If a single normalized character exists, uses that,
otherwise uses combining character.

Also, minor bug fix to similar code in fraction conversion:
as the conversions are made, if there is more than one on
a line, the "end" index becomes stale, since the replacement
text is a different length to the original. Using a mark in
place of an index fixes this.

Fixes #1154