Allow `/r[nn]` to set a right margin
When right-aligning a block, allow the user to specify a
right margin for any block, rather than the default right
or blockquote right margins.

When generating HTML, the right margin on the spans is
adjusted by the amount necessary to allow 0.5em per
character from the default right or blockquote right margins.

It's not appropriate to allow setting of a left margin nor
a first-line left margin on a right-aligned block as is used
on `/#` blocks

Fixes #1145