Avoid `uninitialized` WF error if no spaces in file
Spaces are displayed in the WF Character Count list as `*space*`.
The number of occurrences was set to undef if the file contains
no spaces because of a missing check when space display was
handled specially. Check was there for non-breaking spaces,
but not regular ones. This caused an error when displaying the
character counts.

Also fixed similar (but more minor) error if file contains no
newline characters. Character count displayed a line saying
there were no newlines, instead of omitting mention of
newlines as happens with other characters that aren't in
the file.