Use selected text to populate quicksearch field
If user has selected some text, take the first line only and
put it in the quicksearch entry field, in the same way as the
full S/R dialog.

Also, make Ctrl+Shift+f (the shortcut to pop the dialog) from
within the dialog perform a Search. So user can select text, and
press Ctrl+Shift+f repeatedly to search for selected text.