Add `count` button to Quick Search dialog
Count ("#") button pops dialog with number of occurrences
of string in file. Note that it uses the settings from the Quick
Search dialog, e.g. regex, case, etc.

Shortcut is Ctrl-b, similar to the main S/R dialog, though the
main dialog has a label to show the count as opposed to popping
a quick count dialog. We want to keep the Quick Search dialog
small, so the user can leave it popped up all the time.