Fix S/R in selected range end-point failure
When using Replace All in a selected range, if the replacement
adds a new line, the replacing stopped part way through the
range.

This was caused by the end point being stored as a line number.
The extra line in the replace string caused the actual line number
of the intended end point to increase, so the replacement stopped
early. Fixed by using a mark instead of the line number to note the
end point.

Fixes #1189