Add Illustration Fixup feature to Tools menu (#1196)
Check and report on all illustration markup in downloaded CTF.
Allow user to move illos up/down a paragraph, either to ensure
illo is not mid-paragraph or because illo fits better at the new
location.

Ctrl-click to move illo down. Ctrl-shift-click to move illo up.
Once an illo is moved, Ctrl+shift+j jumps the cursor between original
and new locations for illo.  The "original" location is only
stored for the most recently moved illo, so if you move one
illo, then a different one, then go back to move the first again,
the "original" position of the first is where you left it after its
previous move, not its position when you first loaded the file.
This is implemented with a general "system bookmark"
feature, which could be useful in other places. Menu entry
corresponding to Shift+Ctrl+j is in the Bookmarks menu.

Undo/redo have had to have a slight boost, so that undoing an
illo move can refresh the list.