Implement Sidenote Fixup
Sidenote Fixup is essentially identical to Illustration Fixup, but
it looks for `[Sidenote` instead of `[Illustration`

90% of this commit is just renaming variables and functions, e.g.
`errorcheckilloupdateneeded` to `errorcheckillosnupdatedneeded`
as well as the comments that previously only mentioned illos.

The old Sidenote Fixup and the new one have been moved into
a submenu, and renamed to be Automatic (old) and Manual (new)
versions. The function of the old, automatic fixup is unchanged,
and is still potentially useful to PPers.