Override "busy" behavior for main text window (#1211)
All Tk widgets ring the bell if they receive keystrokes or mouse
clicks while they are busy. However, if the focus changes to the
main text window while the Shift or Control modifier keys are
held down, a "fake" keypress of Shift/Control appears to be
generated, causing the bell to ring.

Override the `_busy` routine to omit the bell if the key stroke
is Shift or Control.