Stop page markers moving in ML Auto-lists
If a multi-line list is converted to HTML using HTML Markup,
Auto List, with the ML flag checked, it deleted all the text and
reinserted a version with `<li>` markup. Unfortunately this
had the effect of moving all the page markers in the list to
the top.

Recoded to work line-by-line, like the non-ML case.

Fixes #1219