Handle page break markers when indenting blocks (#1221)
Code to determine how much to indent/unindent blocks
marked with `/*` did not take note of `x7f` used as a
placeholder for page markers. Hence if a block had a
page break marker at the start of a line,  it miscalculated the
minimum number of leading spaces in the block, and also
did not delete spaces correctly if unindenting was required.