Avoid bug when page marker is on closing markup (#1224)
If there is a blank line or other closing markup immediately
before closing blockquote markup `#/`, the wrapping code
has a special check to avoid accidentally indenting the
closing markup itself.

However, in the previous work (#1111) to ensure the markup
was detected, even if the line was polluted with temporary
page markers, the above case was overlooked. Hence, if there
was a page marker on the closing markup line, the markup
was indented during rewrap.

Using the `notpmselection` variable instead fixes the issue
as expected.