Don't pop Page Marker Adjust when using next page
When using the next/previous page icons on the status bar
(`<` and `>`) it used to pop the Page Marker Adjust dialog,
and turn off Auto Img if that was enabled. It would then
move to the previous/next page marker and show the image,
although it left the cursor just before the page marker, so
in the status bar said you were on the page before the one
that was shown.

Now more consistent. The two buttons move to the previous
or next page marker as before, but only automatically show
the image if Auto Img is enabled; using the buttons does not
turn Auto Img off; the cursor is positioned 1 character into
the correct page, i.e. the earliest point that counts as being
on that page; it also briefly shows the yellow page marker labels,
but doesn't pop the Marker Adjustment dialog.

Unused argument also removed from `togglepagenums()`