Distinguish between classes and elements in pphtml
When checking for unused/undefined CSS, classes and elements
were stored in the same array, which was OK for most things.
However, if a classname was also an element name (e.g. `br`)
it would be incorrectly reported or ignored.

This commit prepends a period to stored class names to
match the class name in the CSS. It also treats anything like
`<string` as a potential element name. This is safer than
trying to maintain a list of all valid element names, and if
the user has an invalid element name, the validator will
tell them, so pphtml doesn't need to do that.