Improve regex checking
User can type regexes in 4 places: Search/Replace, Quick Search,
Word Frequency, Search->Highlight Character, String or Regex.

Only the first had regex checking, changing the text color to red
if the regex gave an error on compilation, and a weak error
message if the user tried to actually search with the bad regex.

This commit improves the error message to report a sanitized
version of the error or warning message for Search/Replace.
It also applies the same color changing, error messages, and
checks for switching between exact/regex as appropriate for
the other 3 cases.

Validating doesn't always work well with having a variable
bound to the text widget, so that has been removed if present,
and the necessary functionality coded separately.