Improve Prev/Next Img status bar button behavior
Historically `<` and `>` in status bar used to cause the Adjust
Page Markers dialog to be popped, and the "previous" and
"next" image moved to were actually the ones before/after
the one displayed in that dialog, regardless of where the cursor
was in the file.
Now that the dialog is not popped, these buttons should go
back/forward a page from the current location. This had a bug,
due to combining `my` with `if`, so location wasn't reset properly,
and back/forward remembered where it had previously gone
even if the cursor had moved.

Behavior now consistent, with only complication being when there
are 2 coincident page markers, when it has to jump over them
both, since there is no location in the file that lies between them.

Fixes #1269