Position coincident page marks in the correct order
If Pg123 and Pg124 were at the same location in the file,
usually because Pg123 is blank, then when the page locations
are loaded from the bin file and positioned in the text widget
using "marks", the mark for Pg124 was inserted at the location
where Pg123's mark was, and ended up before it in the text
widget's internal list of marks.

Loading the marks in reverse order means that coincident
marks appear in alphanumeric order,  which is preferable.

This was assumed to be the case by the HTML generation
code when set to skip coincident pagenum spans. So this
fixes #1272