Sort pagemarks correctly when adding HTML pagenums
Although `markNext` iterates through the marks in order of their
location, it does not guarantee an order when two marks have the
same location. This could lead to some page numbers that directly
followed blank pages not being output to the HTML file.

To resolve it, just sort a list of page markers and use that, rather
than iterating through the markers in the file.

Fixes #1272