Add Index Cross-referencing HTML links
Algorithm searches for "See..." references in the index and
attempts to link them to the relevant index entry. Where it is
unable to find a link due to differences in wording, etc., it
displays a list of errors which can be clicked in order to be
resolved manually.

The Error List has been implemented as a general feature that
other operations could easily use.

Fixes #1264