Support bold/italic page numbers in AutoIndex
If page number was marked bold or italic, the autoindexing code
did not convert it to a hyperlink.

Also, after using a capture group to capture the whole page
number including any bold/italic markup, it was easier to
combine the single page and page range cases into one regex.